requirements.txt
aioredis==2.0.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
black==23.12.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
distlib==0.3.8
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
kombu==5.3.4
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
time-machine==2.13.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
